Refactoring List component to TS
