+++ src/components/List/List.utils.ts
@@ -1,0 +1,4 @@
+export const COMPONENT_KEY = {
+  List: 'List',
+  Item: 'ListItem',
+}
